+++ example/lib/main.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:network_monitoring/network_monitoring.dart';
 import 'package:share_plus/share_plus.dart';
+import 'package:url_launcher/url_launcher.dart';
 
 import 'app.dart';
 import 'config/app_config.dart';
@@ -22,6 +23,10 @@ Future<void> main() async {
     config: NetworkMonitoringConfig(
       shareContent: (context, content) {
         SharePlus.instance.share(ShareParams(text: content));
+      },
+      openUrl: (url) async {
+        final uri = Uri.parse(url);
+        await launchUrl(uri, mode: LaunchMode.externalApplication);
       },
       requiredTaps: AppConfig.devModeRequiredTaps,
       validatePasswordInput: (password) => password == AppConfig.devModePassword,